Moving BindLogDepth to pure file
